+++ bcn_rainfall_core/models/monthly_rainfall.py
@@ -46,6 +46,7 @@ class MonthlyRainfall(YearlyRainfall):
         trace_label: str | None = None,
         plot_average=False,
         plot_linear_regression=False,
+        kmeans_cluster_count: int | None = None,
     ) -> go.Figure | None:
         """
         Overrides parent method by customizing figure and trace labels.
@@ -58,4 +59,5 @@ class MonthlyRainfall(YearlyRainfall):
             trace_label=f"{self.month.value} rainfall",
             plot_average=plot_average,
             plot_linear_regression=plot_linear_regression,
+            kmeans_cluster_count=kmeans_cluster_count,
         )