+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Changelog
 
+### v1.0.4
+_17/06/2025_
+
+- Remove a lot of unused code, notably used to add/remove columns to `DataFrame`.
+- Add function to compute and get K-Means directly for given time frame.
+- Lighten `Label` enum class by removing unused attributes.
+- Integrate display by K-Means clusters in rainfall bar graphic.
+
 ### v1.0.3
 _14/03/2025_